+++ lua/github-theme/palette.lua
@@ -19,14 +19,14 @@ M.setup = function()
   --
 
   -- EndOfBuffer
-  c.sidebar_eob = config.options.dark_sidebar and c.bg2 or c.bg
+  c.sidebar_eob = config.options.darken.sidebars.enable and c.bg2 or c.bg
   c.sidebar_eob = config.options.hide_end_of_buffer and c.sidebar_eob or c.fg_gutter
   c.eob = config.options.hide_end_of_buffer and c.bg or c.fg_gutter
 
   -- Sidebar and Floats
-  c.bg_sidebar = config.options.dark_sidebar and c.bg2 or c.bg
+  c.bg_sidebar = config.options.darken.sidebars.enable and c.bg2 or c.bg
   c.bg_sidebar = config.options.transparent and c.none or c.bg_sidebar
-  c.bg_float = config.options.dark_float and c.bg2 or c.bg
+  c.bg_float = config.options.darken.floats and c.bg2 or c.bg
 
   c = util.color_overrides(c, config.options.colors)
   return c