+++ lua/github-theme/theme.lua
@@ -670,7 +670,7 @@ M.setup = function()
     IndentBlanklineContextStart = { sp = c.fg, underline = true },
   }
 
-  if config.options.hide_inactive_statusline then
+  if config.options.hide_nc_statusline then
     local inactive = { fg = c.bg, bg = c.bg, sp = c.bg_visual, style = Styles.Underline }
     hi.base.StatusLineNC = inactive
     hi.plugins.MiniStatuslineInactive = inactive